feat(game): deal with presence of 'dirty' players in server rooms
Updated the game limit so that the current game can be played with only one player again; also fixed
the situation that players in the game room quit abnormally, but there are still dirty players in
other rooms and on the server.

BREAKING CHANGE: When a player enters someone else's game room, if he/she unexpectedly exits or the
network is disconnected, the other people in the same room and the game room in the server will be
updated to synchronize the latest room status.